ai-everyone/agents/google-agent/requirements.txt
# ARIA Personal AI Assistant — Python Dependencies

# ── Web Framework ─────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# ── Database ──────────────────────────────────────────────
motor==3.5.1          # Async MongoDB driver
pymongo==4.8.0

# ── Authentication & Security ─────────────────────────────
PyJWT==2.9.0
python-jose[cryptography]==3.3.0
passlib==1.7.4

# ── HTTP Client ───────────────────────────────────────────
httpx==0.27.2

# ── Data Validation ───────────────────────────────────────
pydantic==2.9.0
pydantic-settings==2.5.2

# ── Email Handling ────────────────────────────────────────
# (standard library email.mime used — no extra dep needed)

# ── Environment ───────────────────────────────────────────
python-dotenv==1.0.1

# ── Optional: Twilio for calls ────────────────────────────
# twilio==9.3.0

# ── Optional: LangChain integration ──────────────────────
# langchain==0.3.0
# langchain-community==0.3.0
